ai-orchestrator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.11.0
httpx==0.27.0
psycopg2-binary==2.9.11